Move serde serializable types to jormungandr-lib/interfaces Move converting operations to logic
