+++ jormungandr/src/blockchain/reference.rs
@@ -2,8 +2,7 @@ use crate::blockcfg::{
     BlockDate, ChainLength, EpochRewardsInfo, Header, HeaderHash, Leadership, Ledger,
     LedgerParameters,
 };
-use chain_impl_mockchain::certificate::VotePlan;
-use chain_impl_mockchain::multiverse;
+use chain_impl_mockchain::{certificate::VotePlan, multiverse};
 use chain_time::{
     era::{EpochPosition, EpochSlotOffset},
     Epoch, Slot, TimeFrame,
@@ -161,6 +160,10 @@ impl Ref {
         SystemTime::now().duration_since(self.time())
     }
 
+    /// clone all active vote plans at this given state
+    ///
+    /// this includes, votes to be voted on, on going votes, votes to be resolved and votes
+    /// to result into a change on the ledger
     pub fn active_vote_plans(&self) -> Vec<VotePlan> {
         self.ledger.state().active_vote_plans()
     }
